# order-service/docs/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Order service",
    "description": "Service for managing orders",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "OrderService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/orders": {
      "post": {
        "summary": "Создание нового заказа",
        "operationId": "OrderService_CreateOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderserviceCreateOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/orderserviceCreateOrderRequest"
            }
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/api/v1/orders/{orderId}": {
      "get": {
        "summary": "Получение информации о заказе",
        "operationId": "OrderService_GetOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderserviceGetOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    },
    "/api/v1/orders/{orderId}/cancel": {
      "post": {
        "summary": "Отмена заказа",
        "operationId": "OrderService_CancelOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/orderserviceCancelOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "type": "string",
            "format": "int64"
          }
        ],
        "tags": [
          "OrderService"
        ]
      }
    }
  },
  "definitions": {
    "orderserviceCancelOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "Ответ на отмену заказа"
    },
    "orderserviceCreateOrderRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "productId": {
          "type": "string",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "Запрос на создание заказа"
    },
    "orderserviceCreateOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        }
      },
      "title": "Ответ на создание заказа"
    },
    "orderserviceGetOrderResponse": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string",
          "format": "int64"
        },
        "userId": {
          "type": "string",
          "format": "int64"
        },
        "productId": {
          "type": "string",
          "format": "int64"
        },
        "quantity": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "string",
          "format": "int64"
        },
        "status": {
          "$ref": "#/definitions/orderserviceOrderStatus"
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "title": "Ответ с информацией о заказе"
    },
    "orderserviceOrderStatus": {
      "type": "string",
      "enum": [
        "ORDER_STATUS_UNSPECIFIED",
        "ORDER_STATUS_PENDING",
        "ORDER_STATUS_PROCESSING",
        "ORDER_STATUS_APPROVED",
        "ORDER_STATUS_REJECTED",
        "ORDER_STATUS_COMPLETED"
      ],
      "default": "ORDER_STATUS_UNSPECIFIED",
      "title": "Статусы заказа"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}